Smarter shuffle for scry
When adding cards to the top and bottom of the deck via the deck inspector, any cards added to the top or the bottom of the deck will stay in the correct order, even when shuffleing happens on window closed. So you can add a card to the top of your deck, shuffle it when the inspector closes, and it will still be on top.