Fixed new game startup
No streamlined the process for starting a new game on the code side, so everything stays consistent no matter what state the board is in. When you hit new game, you'll always draw 7 cards and the deck will do a shuffle animation. Cards will no longer move and spin accross the entire screen which was fun, but lead to a lot of technical complications and visual artifacts.