Added Jackson Support
Added Jackson library as well as classes to read in card and set json data. This convention will eventually replace the generic JSONObject convention which has to be keyed into manually. In the long-run, this will give us a better hold on our api. Also added code to read and write to file so that we can take the large jsons we initially have when we update and save only the portions of data that we need.