Added Deck Codes
Added buttons to deck editor and collection editor to convert decks to and read decks from a proprietary "deck string". This functionality can help in the case where we can't send a deck directly through Jandor. These deck strings should be able to be copied, pasted, emailed, and used in Jandor to quickly pass a deck around in a pinch. In the future, they may prove useful if there are more programs in the Jandor ecosystem that need to pass decks back and forth.