Refined Encoding
Now using an encoding that maps unique multiverse IDs to indices so we only need to store info on the order of the total number of unique magic cards. This allows us to use a two character sequence for all cards and should be good for up to 32000 cards (Currently there are 20000 or so).